Removed some request and response classes which aren't required as the Reactive client delegates its work to the Daml.Ledger.Client Clients class. Also removed the automation components
