--- src/Daml.Ledger.Api.Data/GetActiveContractsRequest.cs
@@ -1,36 +1,0 @@
-// Copyright(c) 2019 Digital Asset(Switzerland) GmbH and/or its affiliates.All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-namespace Daml.Ledger.Api.Data 
-{
-    using Util;
-
-    public class GetActiveContractsRequest
-    {
-        public GetActiveContractsRequest(string ledgerId, TransactionFilter transactionFilter, bool verbose)
-        {
-            LedgerId = ledgerId;
-            TransactionFilter = transactionFilter;
-            Verbose = verbose;
-        }
-
-        public static GetActiveContractsRequest FromProto(Com.DigitalAsset.Ledger.Api.V1.GetActiveContractsRequest request) => new GetActiveContractsRequest(request.LedgerId, TransactionFilter.FromProto(request.Filter), request.Verbose);
-
-        public Com.DigitalAsset.Ledger.Api.V1.GetActiveContractsRequest ToProto() => new Com.DigitalAsset.Ledger.Api.V1.GetActiveContractsRequest { LedgerId = LedgerId, Filter = TransactionFilter.ToProto(), Verbose = Verbose };
-
-        public string LedgerId { get; }
-
-        public TransactionFilter TransactionFilter {  get; }
-
-        public bool Verbose {  get; }
-
-        public override bool Equals(object obj) => this.Compare(obj, rhs => Verbose == rhs.Verbose && LedgerId == rhs.LedgerId && TransactionFilter == rhs.TransactionFilter);
-
-        public override int GetHashCode() => (LedgerId, TransactionFilter, Verbose).GetHashCode();
-
-        public static bool operator ==(GetActiveContractsRequest lhs, GetActiveContractsRequest rhs) => lhs.Compare(rhs);
-        public static bool operator !=(GetActiveContractsRequest lhs, GetActiveContractsRequest rhs) => !(lhs == rhs);
-
-        public override string ToString() => $"GetActiveContractsRequest{{ledgerId='{LedgerId}', transactionFilter={TransactionFilter}, verbose={Verbose}}}";
-    }
-}
