--- src/Daml.Ledger.Automation/components/helpers/TemplateUtils.cs
@@ -1,10 +1,0 @@
-﻿// Copyright(c) 2019 Digital Asset(Switzerland) GmbH and/or its affiliates.All rights reserved.
-// SPDX-License-Identifier: Apache-2.0
-
-namespace Daml.Ledger.Client.Reactive.Components.Helpers
-{
-    public class TemplateUtils
-    {
-        // Do this later when I look at codegen
-    }
-}
